pyproject.toml
[build-system]
requires = ["hatchling>=1.25"]
build-backend = "hatchling.build"

[project]
name = "revweb"
version = "0.2.0"
description = "CLI + library to crawl and scrape websites into CSV and documentation artifacts for reverse engineering."
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [{name="Your Name", email="you@example.com"}]
keywords = ["scraping", "scrapy", "playwright", "crawler", "reverse-engineering", "cli"]
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
]

dependencies = [
  "typer>=0.12.0",
  "rich>=13.7.0",
  "pydantic>=2.6.0",
  "pyyaml>=6.0.1",
  "python-dateutil>=2.9.0",
  "rapidfuzz>=3.6.0",
  "scrapy>=2.11.0",
  "scrapy-playwright>=0.0.34",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "pytest-cov>=5.0.0",
  "ruff>=0.5.0",
  "mypy>=1.8.0",
  "pre-commit>=3.7.0",
  "types-PyYAML>=6.0.12.20240808",
]

[project.scripts]
revweb = "revweb.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/revweb"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "SIM"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
markers = [
    "unit: Unit tests for individual components",
    "integration: Integration tests for component interactions",
    "e2e: End-to-end tests for CLI and full workflows",
]

[tool.mypy]
python_version = "3.10"
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true
